feat: color enemy and pre-move cannons
